Feat: Interactivity Circle Pointer
