+++ example/lib/main.dart
@@ -1,4 +1,3 @@
-import 'package:example/valuebar_position.dart';
 import 'package:flutter/material.dart';
 import 'package:geekyants_flutter_gauges/geekyants_flutter_gauges.dart';
 
@@ -6,7 +5,7 @@ void main() {
   runApp(
     const MaterialApp(
       debugShowCheckedModeBanner: false,
-      home: MyValueBarPosition(),
+      home: MyGaugeExample(),
     ),
   );
 }
